Add files via upload
Added 'Show Icon For' within Module Settings which facilitates disabling the tokenHUD icon on some tokens. The options are: All - The Icon will appear on every token. Characters - Only appears on Characters. Friendlies - Only on tokens with a Friendly/Neutral disposition. The icon will only ever appear if you're GM, players do not have access to the icon. The default is set to 'Character'.