Add files via upload
Only show icons for Hostiles tokens when Hostile token is hidden to have a cleaner UI and not clutter the TokenHUD

Roll Hostile Stealth (Module Setting) - Button that will roll a stealth check for the selected Hostile token and this is will be the value used against Perception to determine if the hostile token should be revealed.

Display Perception Results (Module Setting) - Print a chat message with each token revealed by which token. i.e. "Donny revealed Shadow [14]".

Account for Walls (Module Settings) - If there's a wall that blocks line of sight, the token will not be revealed.

Distance calculation can now be done using Euclidean math or the grid distance.

Bug Fix: Improved handling of switching hostile tokens to be visible to minimize buggy behavior